refactor: Controller-Duplikate in Trait extrahiert, XMP-Parsing konsolidiert
- StarRateControllerTrait: getUserId(), requireAuth(), getFileById(),
  getJsonBody() aus 4 Controllern extrahiert (18× Auth-Check dedupliziert)
- ExifService: parseXmp() und buildXmpPacket() delegieren an XmpService
- LABEL_MAP kanonisch in XmpService, ExifService referenziert
- ExifServiceTest an neuen Konstruktor (XmpService-Dependency) angepasst
